split course resource
